fix: derive heatmap/volume-profile bucket size from symbol's own price
Both hardcoded a BTC-scale price step (nearest $10, or a flat $0.10/$1
fallback), which collapsed every trade on sub-$1 coins (XRP, DOGE, PEPE,
SHIB, ...) into a single bucket. Added a shared price_step_for() helper
that derives the step from the symbol's current price (fetched once on
connect), snapped to a clean 1/2/5 x 10^n value, so bucket sizing scales
correctly whether the symbol trades at $100k or $0.000009.

+++ backend/app/analytics/heatmap.py
@@ -7,7 +7,7 @@ import time
 from collections import deque
 
 MAX_SNAPSHOTS = 2000
-DEFAULT_STEP  = 10     # price rounding granularity — suitable for BTC ≈ $100 k
+DEFAULT_STEP  = 10.0   # fallback step if the caller can't size one from price
 SNAPSHOT_SECS = 1      # interval between snapshots (seconds)
 
 
@@ -17,14 +17,14 @@ class HeatmapAccumulator:
     Not thread-safe; call exclusively from one asyncio task.
     """
 
-    def __init__(self, price_step: int = DEFAULT_STEP) -> None:
-        self._step      : int              = price_step
-        self._book      : dict[int, float] = {}     # rounded_price → raw_qty
-        self._snapshots : deque[dict]      = deque(maxlen=MAX_SNAPSHOTS)
+    def __init__(self, price_step: float = DEFAULT_STEP) -> None:
+        self._step      : float              = price_step
+        self._book      : dict[float, float] = {}     # rounded_price → raw_qty
+        self._snapshots : deque[dict]        = deque(maxlen=MAX_SNAPSHOTS)
 
     # ── internal ──────────────────────────────────────────────────────────────
 
-    def _round(self, price: float) -> int:
+    def _round(self, price: float) -> float:
         return round(price / self._step) * self._step
 
     def _global_max(self) -> float:
